+++ lib/src/exceptions.dart
@@ -1,18 +1,66 @@
 part of '../msal_js.dart';
 
-AuthException _convertJsAuthError(AuthErrorJs jsError) {
+AuthException _convertJsAuthError(dynamic jsError) {
   if (jsError == null) return null;
 
-  if (jsError is ClientConfigurationErrorJs) {
-    return new ClientConfigurationException._fromJsObject(jsError);
-  } else if (jsError is InteractionRequiredAuthErrorJs) {
-    return new InteractionRequiredAuthException._fromJsObject(jsError);
-  } else if (jsError is ServerErrorJs) {
-    return new ServerException._fromJsObject(jsError);
-  } else if (jsError is ClientAuthErrorJs) {
-    return new ClientAuthException._fromJsObject(jsError);
+  // Convert [jsError] to common type
+  _JsAuthError jsAuthError;
+
+  if (jsError is JsObject) {
+    jsAuthError = _JsAuthError.fromJsObject(jsError);
   } else {
-    return new AuthException._fromJsObject(jsError);
+    jsAuthError = _JsAuthError.fromRawJsObject(jsError);
+  }
+
+  // Determine type
+  final String errorName = jsAuthError.name;
+
+  if (errorName == 'ClientConfigurationError') {
+    return new ClientConfigurationException._fromJsObject(jsAuthError);
+  } else if (errorName == 'InteractionRequiredAuthError') {
+    return new InteractionRequiredAuthException._fromJsObject(jsAuthError);
+  } else if (errorName == 'ServerError') {
+    return new ServerException._fromJsObject(jsAuthError);
+  } else if (errorName == 'ClientAuthError') {
+    return new ClientAuthException._fromJsObject(jsAuthError);
+  } else {
+    return new AuthException._fromJsObject(jsAuthError);
+  }
+}
+
+class _JsAuthError {
+  final String name;
+  final String errorCode;
+  final String errorMessage;
+  final String message;
+  final String stack;
+
+  _JsAuthError._({
+    this.name,
+    this.errorCode,
+    this.errorMessage,
+    this.message,
+    this.stack
+  });
+
+  factory _JsAuthError.fromJsObject(JsObject jsObject) {
+    return _JsAuthError._(
+      name: jsObject['name'],
+      errorCode: jsObject['errorCode'],
+      errorMessage: jsObject['errorMessage'],
+      message: jsObject['message'],
+      stack: jsObject['stack']
+    );
+  }
+  
+  factory _JsAuthError.fromRawJsObject(dynamic jsObject) {
+    return _JsAuthError._(
+      name: getProperty(jsObject, 'name'),
+      errorCode: getProperty(jsObject, 'errorCode'),
+      errorMessage: getProperty(jsObject, 'errorMessage'),
+      message: getProperty(jsObject, 'message'),
+      stack: getProperty(jsObject, 'stack')
+    );
   }
 }
 
@@ -22,12 +70,15 @@ class AuthException implements Exception {
   String get errorCode => _jsObject.errorCode;
   /// A message describing the error.
   String get errorMessage => _jsObject.errorMessage;
+
+  // JS Error fields, AuthError extends Error
+  
   /// Same as [errorMessage].
   String get message => _jsObject.message;
   /// The JavaScript stack trace for the error.
   String get stack => _jsObject.stack;
 
-  final AuthErrorJs _jsObject;
+  final _JsAuthError _jsObject;
 
   AuthException._fromJsObject(this._jsObject);
 
@@ -37,7 +88,7 @@ class AuthException implements Exception {
 
 /// Thrown by MSAL when there is an error in the client code running on the browser.
 class ClientAuthException extends AuthException {
-  ClientAuthException._fromJsObject(ClientAuthErrorJs jsObject)
+  ClientAuthException._fromJsObject(_JsAuthError jsObject)
     : super._fromJsObject(jsObject);
 
   @override
@@ -46,7 +97,7 @@ class ClientAuthException extends AuthException {
 
 /// Thrown by MSAL when there is an error in the configuration of a library object.
 class ClientConfigurationException extends ClientAuthException {
-  ClientConfigurationException._fromJsObject(ClientConfigurationErrorJs jsObject)
+  ClientConfigurationException._fromJsObject(_JsAuthError jsObject)
     : super._fromJsObject(jsObject);
 
   @override
@@ -55,7 +106,7 @@ class ClientConfigurationException extends ClientAuthException {
 
 /// Thrown by MSAL when the user is required to perform an interactive token request.
 class InteractionRequiredAuthException extends ServerException {
-  InteractionRequiredAuthException._fromJsObject(InteractionRequiredAuthErrorJs jsObject)
+  InteractionRequiredAuthException._fromJsObject(_JsAuthError jsObject)
     : super._fromJsObject(jsObject);
 
   @override
@@ -65,7 +116,7 @@ class InteractionRequiredAuthException extends ServerException {
 /// Thrown by MSAL when there is an error with the server code,
 /// for example, unavailability.
 class ServerException extends AuthException {
-  ServerException._fromJsObject(ServerErrorJs jsObject)
+  ServerException._fromJsObject(_JsAuthError jsObject)
     : super._fromJsObject(jsObject);
 
   @override